Added: circle.js, added some basic utils for producing circles Todo: Began updating combatController to allow being in combat and attacking at the same time. Todo: Update AntStateMachine to pull combat states from CombatController
