--- test-fixes.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <title>Quick Test - Ant Controller Fixes</title>
-    <script src="libraries/p5.min.js"></script>
-    <!-- Load all dependencies in correct order -->
-    <script src="Classes/resources.js"></script>
-    <script src="Classes/systems/ResourceManager.js"></script>
-    <script src="Classes/systems/CollisionBox2D.js"></script>
-    <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
-    <script src="Classes/entities/entity.js"></script>
-    <script src="Classes/controllers/MovementController.js"></script>
-    <script src="Classes/controllers/TaskManager.js"></script>
-    <script src="Classes/controllers/RenderController.js"></script>
-    <script src="Classes/controllers/CombatController.js"></script>
-    <script src="Classes/controllers/TransformController.js"></script>
-    <script src="Classes/controllers/SelectionController.js"></script>
-    <script src="Classes/controllers/TerrainController.js"></script>
-    <script src="Classes/controllers/InteractionController.js"></script>
-    <script src="Classes/ants/antStateMachine.js"></script>
-    <script src="Classes/ants/ants.js"></script>
-</head>
-<body>
-    <h1>Ant Controller System Test</h1>
-    <p>Open console (F12) to see test results</p>
-    
-    <script>
-        function runQuickTest() {
-            console.log('🚀 Quick Test: Ant Creation and Controller Access');
-            
-            try {
-                // Debug: Check if controller classes are available
-                console.log('🔍 Debugging controller availability:');
-                console.log('MovementController available:', typeof MovementController !== 'undefined');
-                console.log('TaskManager available:', typeof TaskManager !== 'undefined');
-                console.log('RenderController available:', typeof RenderController !== 'undefined');
-                console.log('Entity available:', typeof Entity !== 'undefined');
-                
-                // Test 1: Basic ant creation
-                console.log('\n📝 Creating ant...');
-                const testAnt = new ant(100, 100, 20, 20, 30, 0, null, "Scout");
-                console.log('✅ Ant created successfully');
-                
-                // Debug: Check Entity controller map
-                console.log('\n🔍 Entity controller debugging:');
-                console.log('_controllers Map size:', testAnt._controllers.size);
-                console.log('Available controllers:');
-                testAnt._controllers.forEach((controller, name) => {
-                    console.log(`  ${name}:`, controller.constructor.name);
-                });
-                
-                // Test 2: Controller access
-                console.log('\n📝 Testing controller access...');
-                console.log('MovementController:', testAnt._movementController ? `✅ ${testAnt._movementController.constructor.name}` : '❌ undefined');
-                console.log('TaskManager:', testAnt._taskManager ? `✅ ${testAnt._taskManager.constructor.name}` : '❌ undefined');
-                console.log('RenderController:', testAnt._renderController ? `✅ ${testAnt._renderController.constructor.name}` : '❌ undefined');
-                
-                // Test 3: Properties
-                console.log('\n📝 Testing properties...');
-                console.log('posX:', testAnt.posX, '✅');
-                console.log('posY:', testAnt.posY, '✅');
-                console.log('speciesName:', testAnt.speciesName, '✅');
-                console.log('isMoving:', testAnt.isMoving, '✅');
-                console.log('center:', testAnt.center, '✅');
-                
-                // Test 4: State machine
-                console.log('\n📝 Testing state machine...');
-                testAnt._stateMachine.update();
-                console.log('✅ State machine update works');
-                
-                // Test 5: Movement (only if movement controller exists)
-                console.log('\n📝 Testing movement...');
-                if (testAnt._movementController) {
-                    const moveResult = testAnt.moveToLocation(200, 150);
-                    console.log('Move result:', moveResult ? '✅' : '❌');
-                } else {
-                    console.log('❌ MovementController not available, skipping movement test');
-                }
-                
-                console.log('\n🎉 Quick test completed!');
-                
-            } catch (error) {
-                console.error('❌ Test failed:', error);
-                console.error('Stack trace:', error.stack);
-            }
-        }
-        
-        // Run test when page loads
-        setTimeout(runQuickTest, 1000);
-    </script>
-</body>
-</html>
