+++ test/README.md
@@ -77,17 +77,17 @@ This test prevents issues like the selection box bug we just fixed, where change
 **Tests Include:**
 - ✅ **Structure Compatibility**: Verifies both original `Ants_Spawn()` and debug command ant creation produce objects with identical structure
 - ✅ **Selection Box Integration**: Tests that selection box functions (`isEntityUnderMouse`, `isEntityInBox`) work with ants created by both methods
-- ✅ **Object Hierarchy Consistency**: Ensures both methods create the same class inheritance chain (AntWrapper → Species → ant)
-- ✅ **Property Accessibility**: Validates that required properties (antObject, species, sprite) are accessible on both object types
+- ✅ **Object Hierarchy Consistency**: Ensures both methods create the same class inheritance chain (AntWrapper → Job → ant)
+- ✅ **Property Accessibility**: Validates that required properties (antObject, Job, sprite) are accessible on both object types
 
 **Key Protections:**
 - Catches when refactoring breaks selection box compatibility
 - Prevents ant creation methods from producing incompatible object structures  
 - Ensures debug tools create ants that work with all game systems
-- Validates Species class inheritance and AntWrapper functionality
+- Validates Job class inheritance and AntWrapper functionality
 
 **When This Test Would Have Caught Our Bug:**
-Our recent fix addressed a case where the debug spawn command created `AntWrapper(ant)` instead of `AntWrapper(Species)`, breaking selection box compatibility. This test would have immediately detected that structural difference.
+Our recent fix addressed a case where the debug spawn command created `AntWrapper(ant)` instead of `AntWrapper(Job)`, breaking selection box compatibility. This test would have immediately detected that structural difference.
 
 ### 🧠 `test/AntStateMachine.test.js`
 **Comprehensive state machine testing**
