+++ test/ant.test.js
@@ -24,7 +24,7 @@ global.devConsoleEnabled = false;
 global.ants = [];
 global.ant_Index = 0;
 
-// Mock stats class
+// Mock StatsContainer class
 class MockStats {
   constructor(position, size, movementSpeed, pendingPos) {
     this.position = { statValue: position };
@@ -33,7 +33,7 @@ class MockStats {
     this.pendingPos = { statValue: pendingPos };
   }
 }
-global.stats = MockStats;
+global.StatsContainer = MockStats;
 
 // Mock Sprite2D class
 class MockSprite2D {
@@ -95,7 +95,7 @@ global.line = () => {};
 class ant {
   constructor(posX = 0, posY = 0, sizex = 50, sizey = 50, movementSpeed = 1, rotation = 0, img = antBaseSprite) {
     const initialPos = createVector(posX, posY);
-    this._stats = new stats(
+    this._stats = new StatsContainer(
       initialPos,
       { x: sizex, y: sizey },
       movementSpeed,
@@ -123,8 +123,8 @@ class ant {
   }
 
   // Getters/Setters
-  get stats() { return this._stats; }
-  set stats(value) { this._stats = value; }
+  get StatsContainer() { return this._stats; }
+  set StatsContainer(value) { this._stats = value; }
   get sprite() { return this._sprite; }
   set sprite(value) { this._sprite = value; }
   get antIndex() { return this._antIndex; }
