+++ test/browser/error-test.html
@@ -59,7 +59,7 @@
                 
                 // Check core classes first
                 checkClass('ant', 'Classes/ants/ants.js');
-                checkClass('Species', 'Classes/ants/species.js');
+                checkClass('Job', 'Classes/ants/Job.js');
                 checkClass('AntStateMachine', 'Classes/ants/antStateMachine.js');
                 
                 // Check controller classes
@@ -118,7 +118,7 @@
     <!-- Load essential scripts in order -->
     <script src="libraries/p5.min.js"></script>
     <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
+    <script src="Classes/entities/StatsContainer.js"></script>
     <script src="Classes/ants/antStateMachine.js"></script>
     <script src="Classes/systems/ResourceManager.js"></script>
     <script src="Classes/managers/GameStateManager.js"></script>
@@ -126,6 +126,6 @@
     <script src="Classes/systems/TaskManager.js"></script>
     <script src="Classes/systems/RenderController.js"></script>
     <script src="Classes/ants/ants.js"></script>
-    <script src="Classes/ants/species.js"></script>
+    <script src="Classes/ants/Job.js"></script>
 </body>
 </html>
