+++ test/browser/integration-status.html
@@ -77,7 +77,7 @@
             // Check if all required classes are available
             const requiredClasses = [
                 'ant', 'MovementController', 'TaskManager', 'RenderController', 
-                'AntStateMachine', 'stats', 'Sprite2D'
+                'AntStateMachine', 'StatsContainer', 'Sprite2D'
             ];
 
             let allAvailable = true;
@@ -312,7 +312,7 @@
     <!-- Load all required scripts -->
     <script src="libraries/p5.min.js"></script>
     <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
+    <script src="Classes/entities/StatsContainer.js"></script>
     <script src="Classes/ants/antStateMachine.js"></script>
     <script src="Classes/systems/ResourceManager.js"></script>
     <script src="Classes/managers/GameStateManager.js"></script>
@@ -320,6 +320,6 @@
     <script src="Classes/systems/TaskManager.js"></script>
     <script src="Classes/systems/RenderController.js"></script>
     <script src="Classes/ants/ants.js"></script>
-    <script src="Classes/ants/species.js"></script>
+    <script src="Classes/ants/Job.js"></script>
 </body>
 </html>
