+++ test/browser/rendercontroller-fix-test.html
@@ -244,7 +244,7 @@
     <!-- Load all required scripts -->
     <script src="libraries/p5.min.js"></script>
     <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
+    <script src="Classes/entities/StatsContainer.js"></script>
     <script src="Classes/ants/antStateMachine.js"></script>
     <script src="Classes/systems/ResourceManager.js"></script>
     <script src="Classes/managers/GameStateManager.js"></script>
@@ -252,6 +252,6 @@
     <script src="Classes/systems/TaskManager.js"></script>
     <script src="Classes/systems/RenderController.js"></script>
     <script src="Classes/ants/ants.js"></script>
-    <script src="Classes/ants/species.js"></script>
+    <script src="Classes/ants/Job.js"></script>
 </body>
 </html>
