+++ test/browser/speed-test.html
@@ -89,7 +89,7 @@
     <!-- Load required scripts -->
     <script src="libraries/p5.min.js"></script>
     <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
+    <script src="Classes/entities/StatsContainer.js"></script>
     <script src="Classes/ants/antStateMachine.js"></script>
     <script src="Classes/systems/ResourceManager.js"></script>
     <script src="Classes/managers/GameStateManager.js"></script>
