+++ test/browser/validation-test.html
@@ -159,7 +159,7 @@
     <!-- Load scripts in correct order -->
     <script src="libraries/p5.min.js"></script>
     <script src="Classes/entities/sprite2d.js"></script>
-    <script src="Classes/entities/stats.js"></script>
+    <script src="Classes/entities/StatsContainer.js"></script>
     <script src="Classes/ants/antStateMachine.js"></script>
     <script src="Classes/systems/ResourceManager.js"></script>
     <script src="Classes/managers/GameStateManager.js"></script>
