+++ test/controllerIntegration.browser.test.js
@@ -137,20 +137,20 @@ function runControllerIntegrationTest() {
     }
   });
   
-  // Test 10: Test Species integration
-  test('Species integration with controllers', () => {
-    // Create ant through Species (like the game does)
+  // Test 10: Test Job integration
+  test('Job integration with controllers', () => {
+    // Create ant through Job (like the game does)
     const baseAnt = new ant(800, 800, 20, 20, 30, 0);
-    const species = new Species(baseAnt, "Builder", antBaseSprite);
+    const Job = new Job(baseAnt, "Builder", antBaseSprite);
     
-    // Species should inherit controller functionality
-    if (!species._movementController) throw new Error('Species missing MovementController');
-    if (!species._taskManager) throw new Error('Species missing TaskManager');
-    if (!species._renderController) throw new Error('Species missing RenderController');
+    // Job should inherit controller functionality
+    if (!Job._movementController) throw new Error('Job missing MovementController');
+    if (!Job._taskManager) throw new Error('Job missing TaskManager');
+    if (!Job._renderController) throw new Error('Job missing RenderController');
     
     // Should be able to move
-    const moveResult = species.moveToLocation(850, 850);
-    if (!moveResult) throw new Error('Species movement failed');
+    const moveResult = Job.moveToLocation(850, 850);
+    if (!moveResult) throw new Error('Job movement failed');
   });
   
   console.log(`\n📊 Integration Test Results: ${testsPassed} passed, ${testsFailed} failed`);
